% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova_by_vars.R
\name{beta_permanova_by_variables}
\alias{beta_permanova_by_variables}
\title{Calculate PERMANOVA results for variables in phyloseq sample_data}
\usage{
beta_permanova_by_variables(
  physeq,
  method = "bray",
  vars = NULL,
  permutations = 999
)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{method}{Distance method for beta diversity (default "bray").}

\item{vars}{Character vector of variable names to test (default all variables in sample_data).}

\item{permutations}{Number of permutations for adonis2 (default 999).}
}
\value{
A tibble with variable names, R2, p-values and permutation count.
}
\description{
Calculate PERMANOVA results for variables in phyloseq sample_data
}
\examples{
data(GlobalPatterns)
res <- beta_permanova_by_variables(GlobalPatterns, method = "bray", permutations = 999)
print(res)
}
