% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phyloseq.R
\name{create_phyloseq}
\alias{create_phyloseq}
\title{Create a phyloseq object from a metdata file and a table containing the abundance and taxonomy data}
\usage{
create_phyloseq(table, metadata, n_sample)
}
\arguments{
\item{table}{A \code{data.frame} containing the abundance matrix in the first \code{n_sample} columns,
followed by taxonomic classification columns.}

\item{metadata}{A \code{data.frame} containing sample metadata. It must either:
\itemize{
\item include a column named "Sample.ID" with sample identifiers (in which case any row names will be removed automatically), or
\item have sample identifiers as row names (in which case they will be converted into a new "Sample.ID" column and row names will be cleared).
}}
}
\value{
A \code{phyloseq} object combining abundance, taxonomy, and sample metadata.
}
\description{
Create a phyloseq object from a metdata file and a table containing the abundance and taxonomy data
}
\examples{
create_phyloseq(main_table, metadata, n_sample = 221)
}
