% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_phy.R
\name{export_phy}
\alias{export_phy}
\title{Export your phyloseq type object as a comma separated value file.}
\usage{
export_phy(phy, path_prefix = "output")
}
\arguments{
\item{phy}{A phyloseq object.}

\item{path_prefix}{The path the file will be saved to.}
}
\value{
A comma separated value file in the specified path
}
\description{
Export your phyloseq type object as a comma separated value file.
}
