% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_taxa.R
\name{get_top_taxa}
\alias{get_top_taxa}
\title{Get the top taxa with most reads in your data}
\usage{
get_top_taxa(physeq, top_n)
}
\arguments{
\item{physeq}{a \code{phyloseq} class object}

\item{top_n}{Integer. The number of top taxa you want to keep.}
}
\value{
A \code{phyloseq} object pruned to keep only the top taxa by abundance.
}
\description{
Get the top taxa with most reads in your data
}
\examples{
get_top_taxa(phy, 50)
}
