% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_plot_from_phy.R
\name{MA_plot_from_phy}
\alias{MA_plot_from_phy}
\title{Create a differential abundance plot using the DESeq2 package from a phyloseq object}
\usage{
MA_plot_from_phy(
  physeq,
  condition_var,
  tax_level = "Genus",
  title = NULL,
  alpha = 0.05,
  return_data = FALSE
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing count data, sample metadata, and taxonomy}

\item{condition_var}{A \code{string} indicating the name of a variable in the sample metadata to compare groups}

\item{tax_level}{A \code{string} indicating the taxonomic level at which to label the taxa (e.g. "Genus", "Family")}

\item{title}{A \code{string} to customize the plot title (optional)}

\item{alpha}{A \code{numeric} value indicating the adjusted p-value threshold for significance (default = 0.05)}

\item{return_data}{Logical; if \code{TRUE}, returns a \code{list} with the plot, full data, and significant taxa}
}
\value{
A ggplot object showing differentially abundant taxa. If \code{return_data = TRUE}, a list with the plot, data, and significant taxa.
}
\description{
Create a differential abundance plot using the DESeq2 package from a phyloseq object
}
\examples{
MA_plot_from_phy(physeq, condition_var = "group", tax_level = "Genus")
}
