% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsummary.R
\name{microsummary}
\alias{microsummary}
\title{Obtain data about the phyloseq dataset}
\usage{
microsummary(x)
}
\arguments{
\item{x}{A phyloseq class Object.}
}
\value{
A tibble containing the total number of reads, OTUs, descriptive statistics of reads, alpha and beta diversity indexes.
}
\description{
Obtain data about the phyloseq dataset
}
\examples{
microsummary(phy)
}
