% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_by_tax.R
\name{plot_bar_by_tax}
\alias{plot_bar_by_tax}
\title{Create composition plots by each taxonomic rank}
\usage{
plot_bar_by_tax(physeq, palette = NULL)
}
\arguments{
\item{physeq}{A \code{phyloseq} class object}

\item{palette}{Optional color vector. If NULL or shorter than the number of taxa,
a palette will be generated automatically.}
}
\value{
A list of \code{ggplot} bar plots, one per taxonomic rank
}
\description{
Generates a list of bar plots for each taxonomic rank found in the
given \code{phyloseq} object. Be mindful that OTU trimming or filtering
might be necessary to visualize the plots adequately.
}
\examples{
plot_bar_by_tax(phy)
}
