% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_by_vars.R
\name{plot_bar_by_vars}
\alias{plot_bar_by_vars}
\title{Create composition plots by each sample variable}
\usage{
plot_bar_by_vars(physeq, rank = "Genus", palette = NULL)
}
\arguments{
\item{physeq}{A \code{phyloseq} class object}

\item{rank}{Character. Taxonomic rank to use for bar plot fill (default is "Genus")}

\item{palette}{Optional color vector. If NULL or shorter than the number of taxa,
a palette will be generated automatically.}
}
\value{
A list of \code{ggplot} bar plots, one per sample variable

A list of \code{ggplot} bar plots, one per sample variable
}
\description{
Generates a list of bar plots grouped by each variable (e.g., sample metadata)
found in the given \code{phyloseq} object. Note that this function assumes
the taxonomic rank "Genus" exists. Plots might not be meaningful for continuous
variables or those with many levels.
}
\examples{
plot_bar_by_vars(phy)
Create composition plots by each sample variable

Generates a list of bar plots grouped by each variable (e.g., sample metadata)
found in the given \code{phyloseq} object. The user can specify the taxonomic rank
used for coloring the bars (e.g., "Phylum", "Genus"). Plots might not be meaningful
for continuous variables or those with many levels.

plot_bar_by_vars(phy, rank = "Phylum")
}
