% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_diversity_by_vars.R
\name{plot_beta_diversity_by_vars}
\alias{plot_beta_diversity_by_vars}
\title{Generate beta diversity ordination plots by sample variables}
\usage{
plot_beta_diversity_by_vars(
  physeq,
  method = "bray",
  ordination_method = "PCoA"
)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{method}{Distance method. Default is "bray". See vegan::vegdist for options.}

\item{ordination_method}{Ordination method. Default is "PCoA".}
}
\value{
A named list of ggplot2 ordination plots.
}
\description{
This function performs ordination on a phyloseq object and generates
plots grouped by variables from sample_data, along with PERMANOVA results.
}
\examples{
data(GlobalPatterns)
beta_plots <- plot_beta_diversity_by_vars(GlobalPatterns)
print(beta_plots[["SampleType"]])
}
