% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwatcher.R
\name{plotwatcher}
\alias{plotwatcher}
\title{Interactive plot visualizer}
\usage{
plotwatcher(plotlist)
}
\arguments{
\item{plotlist}{A list of ggplot2 plot objects.}
}
\value{
Displays plots interactively. Does not return a value.
}
\description{
Allows the user to browse a list of ggplot2 plots either one by one, or all together using patchwork.
}
\examples{
plotwatcher(plotlist)
}
